+++ flake8_async/visitors/visitor91x.py
@@ -18,19 +18,32 @@ from libcst.metadata import PositionProvider
 from ..base import Statement
 from .flake8asyncvisitor import Flake8AsyncVisitor_cst
 from .helpers import (
-    disabled_by_default,
+    AttributeCall,
+    cancel_scope_names,
+    disable_codes_by_default,
     error_class_cst,
+    flatten_preserving_comments,
     fnmatch_qualified_name_cst,
     func_has_decorator,
     iter_guaranteed_once_cst,
+    with_has_call,
 )
 
 if TYPE_CHECKING:
     from collections.abc import Mapping, Sequence
 
 
+class ArtificialStatement(Statement):
+    """Statement that should not trigger 910/911 on function exit.
+
+    Used by loops and `with` statements.
+    """
+
+
 # Statement injected at the start of loops to track missed checkpoints.
-ARTIFICIAL_STATEMENT = Statement("artificial", -1)
+ARTIFICIAL_STATEMENT = ArtificialStatement("artificial", -1)
+# There's no particular reason why loops use a globally instanced statement, but
+# `with` does not - mostly just an artifact of them being implemented at different times.
 
 
 def func_empty_body(node: cst.FunctionDef) -> bool:
@@ -231,8 +244,10 @@ class InsertCheckpointsInLoopBody(CommonVisitors):
     leave_Return = leave_Yield  # type: ignore
 
 
+disable_codes_by_default("ASYNC910", "ASYNC911", "ASYNC912")
+
+
 @error_class_cst
-@disabled_by_default
 class Visitor91X(Flake8AsyncVisitor_cst, CommonVisitors):
     error_codes: Mapping[str, str] = {
         "ASYNC910": (
@@ -242,6 +257,14 @@ class Visitor91X(Flake8AsyncVisitor_cst, CommonVisitors):
         "ASYNC911": (
             "{0} from async iterable with no guaranteed checkpoint since {1.name} "
             "on line {1.lineno}."
+        ),
+        "ASYNC912": (
+            "CancelScope with no guaranteed checkpoint. This makes it potentially "
+            "impossible to cancel."
+        ),
+        "ASYNC100": (
+            "{0}.{1} context contains no checkpoints, remove the context or add"
+            " `await {0}.lowlevel.checkpoint()`."
         ),
     }
 
@@ -256,14 +279,23 @@ class Visitor91X(Flake8AsyncVisitor_cst, CommonVisitors):
         self.loop_state = LoopState()
         self.try_state = TryState()
 
+        # ASYNC100
+        self.has_checkpoint_stack: list[bool] = []
+        self.node_dict: dict[cst.With, list[AttributeCall]] = {}
+
     def should_autofix(self, node: cst.CSTNode, code: str | None = None) -> bool:
+        if code is None:  # pragma: no branch
+            code = "ASYNC911" if self.has_yield else "ASYNC910"
+
         return (
             not self.noautofix
-            and super().should_autofix(
-                node, "ASYNC911" if self.has_yield else "ASYNC910"
-            )
+            and super().should_autofix(node, code)
             and self.library != ("asyncio",)
         )
+
+    def checkpoint(self) -> None:
+        self.uncheckpointed_statements = set()
+        self.has_checkpoint_stack = [True] * len(self.has_checkpoint_stack)
 
     def checkpoint_statement(self) -> cst.SimpleStatementLine:
         return checkpoint_statement(self.library[0])
@@ -283,9 +315,11 @@ class Visitor91X(Flake8AsyncVisitor_cst, CommonVisitors):
             "uncheckpointed_statements",
             "loop_state",
             "try_state",
+            "has_checkpoint_stack",
             copy=True,
         )
         self.uncheckpointed_statements = set()
+        self.has_checkpoint_stack = []
         self.has_yield = self.safe_decorator = False
         self.loop_state = LoopState()
 
@@ -359,7 +393,7 @@ class Visitor91X(Flake8AsyncVisitor_cst, CommonVisitors):
         any_errors = False
         # raise the actual errors
         for statement in self.uncheckpointed_statements:
-            if statement == ARTIFICIAL_STATEMENT:
+            if isinstance(statement, ArtificialStatement):
                 continue
             any_errors |= self.error_91x(original_node, statement)
 
@@ -376,6 +410,7 @@ class Visitor91X(Flake8AsyncVisitor_cst, CommonVisitors):
             self.add_statement = self.checkpoint_statement()
         # avoid duplicate error messages
         self.uncheckpointed_statements = set()
+        # we don't treat it as a checkpoint for ASYNC100
 
         # return original node to avoid problems with identity equality
         assert original_node.deep_equals(updated_node)
@@ -386,7 +421,7 @@ class Visitor91X(Flake8AsyncVisitor_cst, CommonVisitors):
         node: cst.Return | cst.FunctionDef | cst.Yield,
         statement: Statement,
     ) -> bool:
-        assert statement != ARTIFICIAL_STATEMENT
+        assert not isinstance(statement, ArtificialStatement)
 
         if isinstance(node, cst.FunctionDef):
             msg = "exit"
@@ -407,7 +442,7 @@ class Visitor91X(Flake8AsyncVisitor_cst, CommonVisitors):
         # so only set checkpoint after the await node
 
         # all nodes are now checkpointed
-        self.uncheckpointed_statements = set()
+        self.checkpoint()
         return updated_node
 
     # raising exception means we don't need to checkpoint so we can treat it as one
@@ -419,9 +454,54 @@ class Visitor91X(Flake8AsyncVisitor_cst, CommonVisitors):
     # missing-checkpoint warning when there might in fact be one (i.e. a false alarm).
     def visit_With_body(self, node: cst.With):
         if getattr(node, "asynchronous", None):
-            self.uncheckpointed_statements = set()
+            self.checkpoint()
+        if res := (
+            with_has_call(node, *cancel_scope_names)
+            or with_has_call(
+                node, "timeout", "timeout_at", base=("asyncio", "asyncio.timeouts")
+            )
+        ):
+            pos = self.get_metadata(PositionProvider, node).start  # pyright: ignore
+            line: int = pos.line  # pyright: ignore
+            column: int = pos.column  # pyright: ignore
+            self.uncheckpointed_statements.add(
+                ArtificialStatement("with", line, column)
+            )
+            self.node_dict[node] = res
+            self.has_checkpoint_stack.append(False)
+        else:
+            self.has_checkpoint_stack.append(True)
 
-    leave_With_body = visit_With_body
+    def leave_With(self, original_node: cst.With, updated_node: cst.With):
+        # Uses leave_With instead of leave_With_body because we need access to both
+        # original and updated node
+        # ASYNC100
+        if not self.has_checkpoint_stack.pop():
+            autofix = len(updated_node.items) == 1
+            for res in self.node_dict[original_node]:
+                # bypass 910 & 911's should_autofix logic, which excludes asyncio
+                # (TODO: and uses self.noautofix ... which I don't remember what it's for)
+                autofix &= self.error(
+                    res.node, res.base, res.function, error_code="ASYNC100"
+                ) and super().should_autofix(res.node, code="ASYNC100")
+
+            if autofix:
+                return flatten_preserving_comments(updated_node)
+        # ASYNC912
+        else:
+            pos = self.get_metadata(  # pyright: ignore
+                PositionProvider, original_node
+            ).start  # pyright: ignore
+            line: int = pos.line  # pyright: ignore
+            column: int = pos.column  # pyright: ignore
+            s = ArtificialStatement("with", line, column)
+            if s in self.uncheckpointed_statements:
+                self.uncheckpointed_statements.remove(s)
+                for res in self.node_dict[original_node]:
+                    self.error(res.node, error_code="ASYNC912")
+        if getattr(original_node, "asynchronous", None):
+            self.checkpoint()
+        return updated_node
 
     # error if no checkpoint since earlier yield or function entry
     def leave_Yield(
@@ -430,6 +510,10 @@ class Visitor91X(Flake8AsyncVisitor_cst, CommonVisitors):
         if not self.async_function:
             return updated_node
         self.has_yield = True
+
+        # Treat as a checkpoint for ASYNC100, since the context we yield to
+        # may checkpoint.
+        self.has_checkpoint_stack = [True] * len(self.has_checkpoint_stack)
 
         if self.check_function_exit(original_node) and self.should_autofix(
             original_node
@@ -605,7 +689,7 @@ class Visitor91X(Flake8AsyncVisitor_cst, CommonVisitors):
         # appropriate errors if the loop doesn't checkpoint
 
         if getattr(node, "asynchronous", None):
-            self.uncheckpointed_statements = set()
+            self.checkpoint()
         else:
             self.uncheckpointed_statements = {ARTIFICIAL_STATEMENT}
 
@@ -651,7 +735,7 @@ class Visitor91X(Flake8AsyncVisitor_cst, CommonVisitors):
         # AsyncFor guarantees checkpoint on running out of iterable
         # so reset checkpoint state at end of loop. (but not state at break)
         if getattr(node, "asynchronous", None):
-            self.uncheckpointed_statements = set()
+            self.checkpoint()
         else:
             # enter orelse with worst case:
             # loop body might execute fully before entering orelse
@@ -780,7 +864,7 @@ class Visitor91X(Flake8AsyncVisitor_cst, CommonVisitors):
 
         # if async comprehension, checkpoint
         if node.asynchronous:
-            self.uncheckpointed_statements = set()
+            self.checkpoint()
             self.comp_unknown = False
             return False