+++ tests/eval_files/async100_asyncio.py
@@ -1,23 +1,22 @@
 # TRIO_NO_ERROR
 # ANYIO_NO_ERROR
 # BASE_LIBRARY asyncio
-# ASYNCIO_NO_ERROR # TODO
+
+# timeout[_at] re-exported in the main asyncio namespace in py3.11
+# mypy: disable-error-code=attr-defined
+# AUTOFIX
 
 import asyncio
 import asyncio.timeouts
 
 
 async def foo():
-    # py>=3.11 re-exports these in the main asyncio namespace
-    with asyncio.timeout_at(10):  # type: ignore[attr-defined]
+    with asyncio.timeout_at(10):  # error: 9, "asyncio", "timeout_at"
         ...
-    with asyncio.timeout_at(10):  # type: ignore[attr-defined]
+    with asyncio.timeout(10):  # error: 9, "asyncio", "timeout"
         ...
-    with asyncio.timeout(10):  # type: ignore[attr-defined]
+
+    with asyncio.timeouts.timeout_at(10):  # error: 9, "asyncio.timeouts", "timeout_at"
         ...
-    with asyncio.timeouts.timeout_at(10):
-        ...
-    with asyncio.timeouts.timeout_at(10):
-        ...
-    with asyncio.timeouts.timeout(10):
+    with asyncio.timeouts.timeout(10):  # error: 9, "asyncio.timeouts", "timeout"
         ...